# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Undergraduate Hub library and book exchange API
 * OpenAPI spec version: 0.1.0
 */
export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
  rememberSession?: boolean;
}

export type AdminAuthSessionUserRole = typeof AdminAuthSessionUserRole[keyof typeof AdminAuthSessionUserRole];


export const AdminAuthSessionUserRole = {
  staff: 'staff',
  admin: 'admin',
  super_admin: 'super_admin',
} as const;

/**
 * @nullable
 */
export type AdminAuthSessionUser = {
  email: string;
  name: string;
  role: AdminAuthSessionUserRole;
} | null;

export interface AdminAuthSession {
  authenticated: boolean;
  /** @nullable */
  user: AdminAuthSessionUser;
}

export interface HealthStatus {
  status: string;
}

export interface Book {
  id: string;
  title: string;
  author: string;
  category: string;
  language: string;
  /** @nullable */
  coverUrl?: string | null;
  qrCode: string;
  /** @nullable */
  isbn?: string | null;
  price?: number;
  status: string;
  /** @nullable */
  conditionNote?: string | null;
  depositRequired: boolean;
  depositAmount: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  author: string;
  category: string;
  language: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  isbn?: string | null;
  /** @minimum 0 */
  price?: number;
  /** @nullable */
  conditionNote?: string | null;
  depositRequired?: boolean;
  /** @minimum 0 */
  depositAmount?: number;
}

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  author?: string;
  category?: string;
  language?: string;
  /** @minimum 0 */
  price?: number;
  status?: string;
  /** @nullable */
  conditionNote?: string | null;
  depositRequired?: boolean;
  /** @minimum 0 */
  depositAmount?: number;
}

export interface BorrowRequest {
  id: string;
  bookId: string;
  bookTitle: string;
  memberName: string;
  pickupDate: string;
  pickupSlot: string;
  status: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  dueDate?: string | null;
  lateFee: number;
  /** @nullable */
  memberPhone?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  depositStatus?: string | null;
  /** @nullable */
  subscriptionStatus?: string | null;
  /** @nullable */
  approvedPickupDate?: string | null;
  /** @nullable */
  approvedPickupTime?: string | null;
  createdAt: string;
}

export interface BorrowRequestInput {
  bookId: string;
  pickupDate: string;
  pickupSlot: string;
  /** @nullable */
  note?: string | null;
}

export interface RequestStatusUpdate {
  status: string;
  /** @nullable */
  pickupDate?: string | null;
  /** @nullable */
  pickupSlot?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  note?: string | null;
}

export type MemberSubscriptionPlan = typeof MemberSubscriptionPlan[keyof typeof MemberSubscriptionPlan];


export const MemberSubscriptionPlan = {
  NUMBER_29: 29,
  NUMBER_49: 49,
} as const;

export interface Member {
  id: string;
  /** @nullable */
  authUserId?: string | null;
  name: string;
  /** @nullable */
  email: string | null;
  phone: string;
  university?: string;
  /** @nullable */
  studentId?: string | null;
  plan: string;
  subscriptionPlan: MemberSubscriptionPlan;
  /** @nullable */
  subscriptionStart?: string | null;
  /** @nullable */
  subscriptionEnd?: string | null;
  depositAmount?: number;
  depositStatus: string;
  outstandingFees: number;
  status: string;
  totalBooksRead: number;
  joinedAt?: string;
}

export type MemberInputSubscriptionPlan = typeof MemberInputSubscriptionPlan[keyof typeof MemberInputSubscriptionPlan];


export const MemberInputSubscriptionPlan = {
  NUMBER_29: 29,
  NUMBER_49: 49,
} as const;

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 7 */
  phone: string;
  university: string;
  /** @nullable */
  studentId?: string | null;
  subscriptionPlan?: MemberInputSubscriptionPlan;
}

export type MemberUpdateSubscriptionPlan = typeof MemberUpdateSubscriptionPlan[keyof typeof MemberUpdateSubscriptionPlan];


export const MemberUpdateSubscriptionPlan = {
  NUMBER_29: 29,
  NUMBER_49: 49,
} as const;

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 7 */
  phone?: string;
  university?: string;
  /** @nullable */
  studentId?: string | null;
  subscriptionPlan?: MemberUpdateSubscriptionPlan;
  /** @nullable */
  subscriptionStart?: string | null;
  /** @nullable */
  subscriptionEnd?: string | null;
  depositStatus?: string;
  status?: string;
}

export interface Payment {
  id: string;
  memberName: string;
  amount: number;
  method: string;
  type: string;
  status: string;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  transactionId?: string | null;
  paidAt: string;
}

export interface PaymentInput {
  memberId: string;
  /** @minimum 0 */
  amount: number;
  method: string;
  type: string;
  /** @nullable */
  reference?: string | null;
}

export interface WishlistInput {
  bookId: string;
}

export interface MyBook {
  transactionId: string;
  bookId: string;
  title: string;
  author: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  borrowDate?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  returnDate?: string | null;
  status: string;
  lateFee: number;
  daysRemaining: number;
}

export interface Notification {
  id: string;
  title: string;
  message: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}

/**
 * @nullable
 */
export type DashboardActiveBook = {
  title: string;
  author: string;
  dueDate: string;
  daysLeft: number;
  /** @nullable */
  coverUrl?: string | null;
} | null;

export interface Dashboard {
  memberName: string;
  /** @nullable */
  activeBook: DashboardActiveBook;
  dueInDays: number;
  pendingRequests: number;
  wishlistCount: number;
  outstandingFees: number;
  notifications: string[];
}

export type AnalyticsMonthlyBorrowingItem = {
  month: string;
  count: number;
};

export interface Analytics {
  totalBooks: number;
  availableBooks: number;
  borrowedBooks: number;
  activeMembers: number;
  pendingRequests: number;
  monthlyBorrowing: AnalyticsMonthlyBorrowingItem[];
}

export type ListBooksParams = {
search?: string;
category?: string;
language?: string;
status?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 48
 */
pageSize?: number;
sort?: ListBooksSort;
};

export type ListBooksSort = typeof ListBooksSort[keyof typeof ListBooksSort];


export const ListBooksSort = {
  title: 'title',
  recent: 'recent',
  popular: 'popular',
} as const;