# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Undergraduate Hub library and book exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Sign in an administrator through Supabase Auth
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLoginInput" }
      responses:
        "200":
          description: Authenticated admin session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAuthSession" }
        "401": { description: Invalid credentials }
        "403": { description: Authenticated account is not an administrator }
        "503": { description: Admin authentication is not configured }
...
components:
  schemas:
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
        rememberSession: { type: boolean, default: false }
    AdminAuthSession:
      type: object
      required: [authenticated, user]
      properties:
        authenticated: { type: boolean }
        user:
          type: ["object", "null"]
          properties:
            email: { type: string }
            name: { type: string }
            role: { type: string, enum: [staff, admin, super_admin] }
          required: [email, name, role]
...