# lib/api-zod/src/generated/types/adminAuthSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Undergraduate Hub library and book exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAuthSessionUser } from './adminAuthSessionUser';

export interface AdminAuthSession {
  authenticated: boolean;
  /** @nullable */
  user: AdminAuthSessionUser;
}