# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Undergraduate Hub library and book exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
  rememberSession?: boolean;
}